+++ R/find_countrycol.R
@@ -10,7 +10,7 @@
 #' @param min_share A value between \code{0} and \code{1} indicating the minimum share of country names in columns that are returned. A value of \code{0} will return any column containing a country name. A value of \code{1} will return only columns whose entries are all country names. Default is \code{0.9}, i.e. at least 90 percent of the column entries need to be country names.
 #' @param sample_size A numeric value greater than 0 indicating the sample size used for evaluating columns in large datasets.  Default is \code{1000}. Large sample sizes result in slower execution speed.
 #' @return Returns a vector of country names (\code{return_index=FALSE}) or column indices (\code{return_index=TRUE}) of columns containing country names.
-#' @seealso \link[countries]{is_country} \link[countries]{country_name} \link[countries]{find_timecol}
+#' @seealso \link[countries]{is_country} \link[countries]{country_name} \link[countries]{find_keycol} \link[countries]{find_timecol}
 #' @export
 #' @examples
 #' find_countrycol(x=data.frame(a=c("Brésil","Tonga","FRA"), b=c(1,2,3)))