+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added function *is_keycol()* to test whether a set of columns could be the keys of a table
 * Added function *find_countrycol()* to automate search of columns containing country names in data frames
 * Added function *find_timecol()* to automate search of date and/or year columns in data frames
+* Added function *find_keycol()* to automate search of table keys in data frames
 * Speed improvements for the functions *country_name()* and *match_table()*
 * Bug fixes